Integrate `to_bounded_ordered_map` with `to_ordered_map`
Control top/tail-k or simple natural/reverse ordered map via 3rd
argument
